apps/entity-service/src/trees/research.service.ts
import { TreeSpeciesResearch } from "@terramatch-microservices/database/entities";
import { Op } from "sequelize";
import { Injectable } from "@nestjs/common";

@Injectable()
export class ResearchService {
  async searchScientificNames(search: string) {
    return (
      await TreeSpeciesResearch.findAll({
        where: {
          [Op.or]: [
            // By checking these two, we're limiting the search term to only occurrences at the
            // beginning of a word in the scientific name, which tends to lead to better results.
            { scientificName: { [Op.like]: `${search}%` } },
            { scientificName: { [Op.like]: `% ${search}%` } }
          ]
        },
        attributes: ["taxonId", "scientificName"],
        limit: 10
      })
    ).map(({ taxonId, scientificName }) => ({ taxonId, scientificName }));
  }
}
